Move paired device trust to agents
