+++ apps/dd-agent/workload.json.tmpl
@@ -28,6 +28,7 @@
     "DD_CONFIDENTIAL=${DD_CONFIDENTIAL}",
     "DD_PORT=8080",
     "DD_EXTRA_INGRESS=${DD_EXTRA_INGRESS}",
+    "DD_AGENT_DEVICES_PATH=/var/lib/easyenclave/data/dd-agent/devices.json",
     "DD_ORACLES_B64=${DD_ORACLES_B64}"
   ]
 }
